Add saving now optional, Fix crop! error for element out of viewport
Added option to save the screenshot or just to recieve it back as
ChunkyPNG::Image so users are now able to work with the image in
their automated test.

Fixed error of crop! when the element was too big for the viewport
or when it was somewhere in the middle of the page.

Know problem - You still need to scroll to the element and it needs
to be visible on the page.